refactor(layout): change layout checksum type size

+++ src/variables/layout/layout_checksum.rs
@@ -1,14 +1,15 @@
 pub struct LayoutChecksum;
 
 impl LayoutChecksum {
-    // NOTE: source layout-custom.c
-    pub fn calc(s: &str) -> usize {
-        let mut csum: usize = 0;
+    // NOTE: tmux source -> layout-custom.c
+    // BSD checksum algorithm (without in bounds AND 0b1111_1111)
+    pub fn calc(s: &str) -> u16 {
+        let mut csum: u16 = 0;
 
         let chars = s.chars();
         for chr in chars {
             csum = (csum >> 1) + ((csum & 1) << 15);
-            csum += chr as usize;
+            csum += chr as u16;
         }
         csum
     }
